+++ apps/www/src/app/(main)/(docs)/components/react-hook-form/examples/tags-input.tsx
@@ -3,22 +3,22 @@
 import { zodResolver } from "@hookform/resolvers/zod";
 import { RHFTagsInput } from "@jamsr-ui/react";
 import { useForm } from "react-hook-form";
-import { object, set, string } from "zod";
+import { array, object, string } from "zod";
 import { RHFDemoWrapper } from "../components/wrapper";
 
 type FormValues = {
-  tags: Set<string>;
+  tags: string[];
 };
 
 const schema = object({
-  tags: set(string().trim().min(2, "Tags should be atleast 2 characters"))
+  tags: array(string().trim().min(2, "Tags should be atleast 2 characters"))
     .min(2, "Minimum 2 tags required")
     .max(10, "Maximum 10 tags allowed"),
 });
 
 export const RHFDemoTagsInput = () => {
   const defaultValues: FormValues = {
-    tags: new Set([]),
+    tags: [],
   };
   const methods = useForm<FormValues>({
     defaultValues,