+++ packages/components/tags-input/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @jamsr-ui/tags-input
 
+## 0.0.29
+
+### Patch Changes
+
+- fix: tags-input
+
 ## 0.0.28
 
 ### Patch Changes